backend/requirements.txt
Django==3.2.23
django-cors-headers==4.1.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
certifi==2023.11.17
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
colorama==0.4.6
executing==2.0.1
icecream==2.1.3
idna==3.6
importlib-metadata==6.7.0
msgpack==1.0.5
passlib==1.7.4
Pillow==9.5.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pyotp==2.9.0
pypng==0.20220715.0
pytz==2023.3.post1
qrcode==7.4.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
python-magic==0.4.27